requirements.txt
google-cloud-texttospeech>=2.14.0
google-cloud-storage>=2.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
